+++ app/services/import_planned_disbursements.rb
@@ -111,6 +111,8 @@ class ImportPlannedDisbursements
   rescue Encoding::CompatibilityError
     value.force_encoding(Encoding::UTF_8)
     @errors << Error.new(@current_index, header, value, I18n.t("importer.errors.planned_disbursement.invalid_characters"))
+  rescue PlannedDisbursementHistory::SequenceError
+    @errors << Error.new(@current_index, header, header, I18n.t("importer.errors.planned_disbursement.in_the_past"))
   end
 
   def lookup_activity(roda_identifier)