+++ spec/features/staff/beis_users_can_edit_a_transfer_spec.rb
@@ -2,8 +2,8 @@ RSpec.feature "BEIS users can edit a transfer" do
   let(:user) { create(:beis_user) }
   before { authenticate!(user: user) }
 
-  let(:source_activity) { create(:activity) }
-  let(:destination_activity) { create(:activity) }
+  let(:source_activity) { create(:project_activity) }
+  let(:destination_activity) { create(:project_activity) }
 
   let!(:transfer) { create(:transfer, source: source_activity, destination: destination_activity) }
 
@@ -65,7 +65,7 @@ RSpec.feature "BEIS users can edit a transfer" do
   end
 
   scenario "the user can see validation errors" do
-    non_existent_activity = build(:activity)
+    non_existent_activity = build(:project_activity)
 
     roda_identifier = "GCRF-BLOB-424434434"
     allow(non_existent_activity).to receive(:roda_identifier) { roda_identifier }