+++ spec/models/activity_spec.rb
@@ -91,14 +91,14 @@ RSpec.describe Activity, type: :model do
     describe ".projects_and_third_party_projects_for_report" do
       it "only returns projects and third party projects that are for the reports organisation and fund" do
         organisation = create(:delivery_partner_organisation)
-        fund = create(:fund_activity, organisation: organisation)
-        programme = create(:programme_activity, parent: fund, organisation: organisation)
+        fund = create(:fund_activity)
+        programme = create(:programme_activity, parent: fund)
         project = create(:project_activity, parent: programme, organisation: organisation)
         third_party_project = create(:third_party_project_activity, parent: project, organisation: organisation)
         report = create(:report, organisation: third_party_project.organisation, fund: fund)
 
-        another_fund = create(:fund_activity, organisation: organisation)
-        another_programme = create(:programme_activity, parent: another_fund, organisation: organisation)
+        another_fund = create(:fund_activity)
+        another_programme = create(:programme_activity, parent: another_fund)
         another_project = create(:project_activity, parent: another_programme, organisation: organisation)
         another_third_party_project = create(:third_party_project_activity, parent: another_project, organisation: organisation)
 
@@ -129,28 +129,28 @@ RSpec.describe Activity, type: :model do
 
     context "overall activity state" do
       context "when the activity form is final" do
-        subject { build(:activity, :at_identifier_step, form_state: "complete") }
+        subject { build(:project_activity, :at_identifier_step, form_state: "complete") }
         it { should be_invalid }
       end
     end
 
     context "#form_state" do
       context "when the form_state is set to a value we expect" do
-        subject(:activity) { build(:activity) }
+        subject(:activity) { build(:project_activity) }
         it "should be valid" do
           expect(activity.valid?).to be_truthy
         end
       end
 
       context "when form_state is set to a value not included in the validation list" do
-        subject(:activity) { build(:activity, form_state: "completed") }
+        subject(:activity) { build(:project_activity, form_state: "completed") }
         it "should not be valid" do
           expect(activity.valid?).to be_falsey
         end
       end
 
       context "when form_state is set to a value included in the validation list" do
-        subject(:activity) { build(:activity, form_state: "purpose") }
+        subject(:activity) { build(:project_activity, form_state: "purpose") }
         it "should be valid" do
           expect(activity.valid?).to be_truthy
         end
@@ -158,7 +158,7 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when delivery_partner_identifier is blank" do
-      subject(:activity) { build(:activity, delivery_partner_identifier: nil) }
+      subject(:activity) { build(:project_activity, delivery_partner_identifier: nil) }
       it "should not be valid" do
         expect(activity.valid?(:identifier_step)).to be_falsey
       end
@@ -349,14 +349,14 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when title is blank" do
-      subject(:activity) { build(:activity, title: nil) }
+      subject(:activity) { build(:project_activity, title: nil) }
       it "should not be valid" do
         expect(activity.valid?(:purpose_step)).to be_falsey
       end
     end
 
     context "when description is blank" do
-      subject(:activity) { build(:activity, description: nil) }
+      subject(:activity) { build(:project_activity, description: nil) }
       it "should not be valid" do
         expect(activity.valid?(:purpose_step)).to be_falsey
       end
@@ -370,73 +370,73 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when sector category is blank" do
-      subject(:activity) { build(:activity, sector_category: nil) }
+      subject(:activity) { build(:project_activity, sector_category: nil) }
       it "should not be valid" do
         expect(activity.valid?(:sector_category_step)).to be_falsey
       end
     end
 
     context "when sector is blank" do
-      subject(:activity) { build(:activity, sector: nil) }
+      subject(:activity) { build(:project_activity, sector: nil) }
       it "should not be valid" do
         expect(activity.valid?(:sector_step)).to be_falsey
       end
     end
 
     context "when planned start and actual start dates are blank" do
-      subject(:activity) { build(:activity, planned_start_date: nil, actual_start_date: nil) }
+      subject(:activity) { build(:project_activity, planned_start_date: nil, actual_start_date: nil) }
       it "should not be valid" do
         expect(activity.valid?(:dates_step)).to be_falsey
       end
     end
 
     context "when programme status is blank" do
-      subject(:activity) { build(:activity, programme_status: nil) }
+      subject(:activity) { build(:project_activity, programme_status: nil) }
       it "should not be valid" do
         expect(activity.valid?(:programme_status_step)).to be_falsey
       end
     end
 
     context "when planned_start_date is blank but actual_start_date is not nil" do
-      subject(:activity) { build(:activity, planned_start_date: nil) }
+      subject(:activity) { build(:project_activity, planned_start_date: nil) }
       it "should be valid" do
         expect(activity.valid?(:dates_step)).to be_truthy
       end
     end
 
     context "when actual_start_date is blank but planned_start_date is not nil" do
-      subject(:activity) { build(:activity, actual_start_date: nil) }
+      subject(:activity) { build(:project_activity, actual_start_date: nil) }
       it "should be valid" do
         expect(activity.valid?(:dates_step)).to be_truthy
       end
     end
 
     context "when planned_end_date is blank" do
-      subject(:activity) { build(:activity, planned_end_date: nil) }
+      subject(:activity) { build(:project_activity, planned_end_date: nil) }
       it "should be valid" do
         expect(activity.valid?(:dates_step)).to be_truthy
       end
     end
 
     context "when actual_start_date is blank" do
-      subject(:activity) { build(:activity, actual_start_date: nil) }
+      subject(:activity) { build(:project_activity, actual_start_date: nil) }
       it "should be valid" do
         expect(activity.valid?(:dates_step)).to be_truthy
       end
     end
 
     context "when actual_end_date is blank" do
-      subject(:activity) { build(:activity, actual_end_date: nil) }
+      subject(:activity) { build(:project_activity, actual_end_date: nil) }
       it "should be valid" do
         expect(activity.valid?(:dates_step)).to be_truthy
       end
     end
 
     context "when planned_end_date is not blank" do
-      let(:activity) { build(:activity) }
+      let(:activity) { build(:project_activity) }
 
       it "does not allow planned_end_date to be earlier than planned_start_date" do
-        activity = build(:activity, planned_start_date: Date.today, planned_end_date: Date.yesterday)
+        activity = build(:project_activity, planned_start_date: Date.today, planned_end_date: Date.yesterday)
         expect(activity.valid?).to be_falsey
         expect(activity.errors[:planned_end_date]).to include "Planned end date must be after planned start date"
       end
@@ -467,7 +467,7 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when geography is blank" do
-      subject(:activity) { build(:activity, geography: nil) }
+      subject(:activity) { build(:project_activity, geography: nil) }
       it "should not be valid" do
         expect(activity.valid?(:geography_step)).to be_falsey
       end
@@ -475,7 +475,7 @@ RSpec.describe Activity, type: :model do
 
     context "when geography is recipient_region" do
       context "and recipient_region and recipient_contry are blank" do
-        subject { build(:activity) }
+        subject { build(:project_activity) }
         it { should validate_presence_of(:recipient_region).on(:region_step) }
         it { should_not validate_presence_of(:recipient_country).on(:country_step) }
       end
@@ -483,28 +483,28 @@ RSpec.describe Activity, type: :model do
 
     context "when geography is recipient_country" do
       context "and recipient_region and recipient_country are blank" do
-        subject { build(:activity, geography: :recipient_country) }
+        subject { build(:project_activity, geography: :recipient_country) }
         it { should validate_presence_of(:recipient_country).on(:country_step) }
         it { should_not validate_presence_of(:recipient_region).on(:region_step) }
       end
     end
 
     context "when requires_additional_benefitting_countries is blank when required" do
-      subject(:activity) { build(:activity, geography: :recipient_country, requires_additional_benefitting_countries: nil) }
+      subject(:activity) { build(:project_activity, geography: :recipient_country, requires_additional_benefitting_countries: nil) }
       it "should not be valid" do
         expect(activity.valid?(:requires_additional_benefitting_countries_step)).to be_falsey
       end
     end
 
     context "when intended_beneficiaries is blank" do
-      subject(:activity) { build(:activity, intended_beneficiaries: nil) }
+      subject(:activity) { build(:project_activity, intended_beneficiaries: nil) }
       it "should not be valid" do
         expect(activity.valid?(:intended_beneficiaries_step)).to be_falsey
       end
     end
 
     context "when gdi is blank" do
-      subject(:activity) { build(:activity, gdi: nil) }
+      subject(:activity) { build(:project_activity, gdi: nil) }
       it "should not be valid" do
         expect(activity.valid?(:gdi_step)).to be_falsey
       end
@@ -548,7 +548,7 @@ RSpec.describe Activity, type: :model do
       end
 
       it "is not required if the activity is a fund" do
-        activity = build(:activity, level: :fund, fund_pillar: nil)
+        activity = build(:fund_activity, fund_pillar: nil)
 
         expect(activity.valid?(:fund_pillar_step)).to be_truthy
       end
@@ -569,14 +569,14 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when fstc applies is blank" do
-      subject(:activity) { build(:activity, fstc_applies: nil) }
+      subject(:activity) { build(:project_activity, fstc_applies: nil) }
       it "should not be valid" do
         expect(activity.valid?(:fstc_applies_step)).to be_falsey
       end
     end
 
     context "when Covid19-related research is blank" do
-      subject(:activity) { build(:activity, covid19_related: nil) }
+      subject(:activity) { build(:project_activity, covid19_related: nil) }
       it "should not be valid" do
         expect(activity.valid?(:covid19_related_step)).to be_falsey
       end
@@ -650,7 +650,7 @@ RSpec.describe Activity, type: :model do
     end
 
     context "when oda_eligibility is blank" do
-      subject(:activity) { build(:activity, oda_eligibility: nil) }
+      subject(:activity) { build(:project_activity, oda_eligibility: nil) }
       it "should not be valid" do
         expect(activity.valid?(:oda_eligibility_step)).to be_falsey
       end
@@ -658,34 +658,34 @@ RSpec.describe Activity, type: :model do
 
     context "when saving in the oda_eligibility_lead_step context" do
       context "and the activity is a fund" do
-        subject { build(:activity, level: :fund) }
+        subject { build(:project_activity, level: :fund) }
         it { should_not validate_presence_of(:oda_eligibility_lead).on(:oda_eligibility_lead_step) }
       end
 
       context "and the activity is a programme" do
-        subject { build(:activity, level: :programme) }
+        subject { build(:project_activity, level: :programme) }
         it { should_not validate_presence_of(:oda_eligibility_lead).on(:oda_eligibility_lead_step) }
       end
 
       context "and the activity is a project" do
-        subject { build(:activity, level: :project) }
+        subject { build(:project_activity, level: :project) }
         it { should validate_presence_of(:oda_eligibility_lead).on(:oda_eligibility_lead_step) }
       end
 
       context "and the activity is a third party project" do
-        subject { build(:activity, level: :third_party_project) }
+        subject { build(:project_activity, level: :third_party_project) }
         it { should validate_presence_of(:oda_eligibility_lead).on(:oda_eligibility_lead_step) }
       end
     end
 
     context "when saving in the uk_dp_named_contact context" do
       context "and the activity is a fund" do
-        subject { build(:activity, level: :fund) }
+        subject { build(:project_activity, level: :fund) }
         it { should_not validate_presence_of(:uk_dp_named_contact).on(:uk_dp_named_contact_step) }
       end
 
       context "and the activity is a programme" do
-        subject { build(:activity, level: :programme) }
+        subject { build(:project_activity, level: :programme) }
         it { should_not validate_presence_of(:uk_dp_named_contact).on(:uk_dp_named_contact_step) }
       end
 
@@ -763,9 +763,11 @@ RSpec.describe Activity, type: :model do
     end
 
     describe "parent association" do
-      subject { Activity.new(level: level) }
+      let(:organisation) { build(:delivery_partner_organisation) }
+      subject { Activity.new(level: level, organisation: organisation) }
 
       context "with a fund" do
+        let(:organisation) { build(:beis_organisation) }
         let(:level) { "fund" }
 
         it { is_expected.to validate_absence_of :parent }
@@ -894,19 +896,19 @@ RSpec.describe Activity, type: :model do
 
   describe "#form_steps_completed?" do
     it "is true when a user has completed all of the form steps" do
-      activity = build(:activity, form_state: :complete)
+      activity = build(:project_activity, form_state: :complete)
 
       expect(activity.form_steps_completed?).to be_truthy
     end
 
     it "is false when a user is still completing one of the form steps" do
-      activity = build(:activity, form_state: :purpose)
+      activity = build(:project_activity, form_state: :purpose)
 
       expect(activity.form_steps_completed?).to be_falsey
     end
 
     it "is false when the form_state is nil" do
-      activity = build(:activity, form_state: nil)
+      activity = build(:project_activity, form_state: nil)
 
       expect(activity.form_steps_completed?).to be_falsey
     end
@@ -921,7 +923,7 @@ RSpec.describe Activity, type: :model do
   end
 
   it "returns false if all extending_organisation fields are not present" do
-    activity = build(:activity)
+    activity = build(:project_activity, extending_organisation: nil)
 
     expect(activity.has_extending_organisation?).to be false
   end
@@ -1215,12 +1217,12 @@ RSpec.describe Activity, type: :model do
 
   describe "#iati_identifier" do
     it "returns the previous_identifier if it exists" do
-      activity = create(:activity, previous_identifier: "previous-id", transparency_identifier: "transparency-id")
+      activity = create(:project_activity, previous_identifier: "previous-id", transparency_identifier: "transparency-id")
       expect(activity.iati_identifier).to eq("previous-id")
     end
 
     it "returns the transparency_identifier if previous_identifier is not set" do
-      activity = create(:activity, previous_identifier: nil, transparency_identifier: "transparency-id")
+      activity = create(:project_activity, previous_identifier: nil, transparency_identifier: "transparency-id")
       expect(activity.iati_identifier).to eq("transparency-id")
     end
   end
@@ -1518,7 +1520,7 @@ RSpec.describe Activity, type: :model do
 
   describe "#source_fund" do
     context "for a Newton fund activity" do
-      let(:activity) { build(:activity, source_fund_code: Fund::MAPPINGS["NF"]) }
+      let(:activity) { build(:project_activity, source_fund_code: Fund::MAPPINGS["NF"]) }
 
       it "returns a Newton fund" do
         expect(activity.source_fund).to be_a(Fund)
@@ -1528,7 +1530,7 @@ RSpec.describe Activity, type: :model do
     end
 
     context "for a GCRF activity" do
-      let(:activity) { build(:activity, source_fund_code: Fund::MAPPINGS["GCRF"]) }
+      let(:activity) { build(:project_activity, source_fund_code: Fund::MAPPINGS["GCRF"]) }
 
       it "returns a GCRF fund" do
         expect(activity.source_fund).to be_a(Fund)
@@ -1540,7 +1542,7 @@ RSpec.describe Activity, type: :model do
 
   describe "#source_fund=" do
     it "sets the source fund code" do
-      activity = build(:activity)
+      activity = build(:project_activity)
       activity.source_fund = Fund.new(Fund::MAPPINGS["GCRF"])
       activity.save