+++ spec/validators/organisation_validator_spec.rb
@@ -1,0 +1,67 @@
+require "rails_helper"
+
+RSpec.describe OrganisationValidator do
+  context "when activity is a fund" do
+    subject { build(:fund_activity, organisation: organisation) }
+
+    context "when the organisation is a delivery partner" do
+      let(:organisation) { build(:delivery_partner_organisation) }
+
+      it { should be_invalid }
+    end
+
+    context "when the organisation is the service owner" do
+      let(:organisation) { build(:beis_organisation) }
+
+      it { should be_valid }
+    end
+  end
+
+  context "when activity is a programme" do
+    subject { build(:fund_activity, organisation: organisation) }
+
+    context "when the organisation is a delivery partner" do
+      let(:organisation) { build(:delivery_partner_organisation) }
+
+      it { should be_invalid }
+    end
+
+    context "when the organisation is the service owner" do
+      let(:organisation) { build(:beis_organisation) }
+
+      it { should be_valid }
+    end
+  end
+
+  context "when activity is a project" do
+    subject { build(:project_activity, organisation: organisation) }
+
+    context "when the organisation is a delivery partner" do
+      let(:organisation) { build(:delivery_partner_organisation) }
+
+      it { should be_valid }
+    end
+
+    context "when the organisation is the service owner" do
+      let(:organisation) { build(:beis_organisation) }
+
+      it { should be_invalid }
+    end
+  end
+
+  context "when activity is a third party project" do
+    subject { build(:third_party_project_activity, organisation: organisation) }
+
+    context "when the organisation is a delivery partner" do
+      let(:organisation) { build(:delivery_partner_organisation) }
+
+      it { should be_valid }
+    end
+
+    context "when the organisation is the service owner" do
+      let(:organisation) { build(:beis_organisation) }
+
+      it { should be_invalid }
+    end
+  end
+end